+++ servant-routes/src/Servant/API/Routes/Utils.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE PolyKinds #-}
+
 {- |
 Module      : Servant.API.Routes.Utils
 Copyright   : (c) Frederick Pringle, 2025